# src/main/webapp/swagger/index.yaml
swagger: '2.0'
info:
  description: JAVA后台管理系统，企业开发首选脚手架。
  version: 1.0.0
  title: SpringAdmin
basePath: /SpringAdmin
schemes:
  - http

#定义接口数据
paths:
...
  /sys/user/list:
    get:
      tags:
        - 用户管理
      summary: 查询用户列表
      produces:
        - application/json
      parameters:
        - name: pageSize
          description: 分页条数
          in: query
          type: integer
          required: true
        - name: pageNumber
          description: 分页页码
          in: query
          type: integer
          required: true
        - name: sortName
          description: 排序字段
          in: query
          type: string
          required: true
        - name: sortOrder
          description: 排序方式，如：asc、desc
          in: query
          type: string
          required: true
        - name: username
          description: 用户名
          in: query
          type: string
      responses:
        '200':
          description: 返回用户列表
          schema:
            $ref: '#/definitions/SysUserInfo'
  /sys/user/save:
    post:
      tags:
        - 用户管理
      summary: 新增用户
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserInfo'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/user/update:
    post:
      tags:
        - 用户管理
      summary: 修改用户
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserInfo'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/user/delete:
    post:
      tags:
        - 用户管理
      summary: 删除用户
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/user/updatePwd:
      post:
        tags:
          - 用户管理
        summary: 修改密码
        produces:
          - application/json
        parameters:
          - name: oldPwd
            description: 原密码
            in: query
            type: string
            required: true
          - name: newPwd
            description: 新密码
            in: query
            type: string
            required: true
        responses:
          '200':
            schema:
              $ref: '#/definitions/Result'
...
  /sys/role/list:
    get:
      tags:
        - 角色管理
      summary: 查询角色列表
      produces:
        - application/json
      parameters:
        - name: pageSize
          description: 分页条数
          in: query
          type: integer
          required: true
        - name: pageNumber
          description: 分页页码
          in: query
          type: integer
          required: true
        - name: sortName
          description: 排序字段
          in: query
          type: string
          required: true
        - name: sortOrder
          description: 排序方式，如：asc、desc
          in: query
          type: string
          required: true
        - name: roleName
          description: 角色名
          in: query
          type: string
      responses:
        '200':
          description: 角色列表
          schema:
            $ref: '#/definitions/SysRoleInfo'
...
  /sys/role/save:
    post:
      tags:
        - 角色管理
      summary: 新增角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleInfo'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/role/update:
    post:
      tags:
        - 角色管理
      summary: 修改角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleInfo'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/role/delete:
    post:
      tags:
        - 角色管理
      summary: 删除角色
      produces:
        - application/json
      parameters:
        - name: list
          description: 角色ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
...
definitions:
  Result:
      type: object
      properties:
        code:
          description: 状态码 1：成功  0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
  SysUserInfo:
    type: object
    properties:
      id:
        description: 用户ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      realname:
        description: 用户姓名
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      password:
        description: 密码
        type: string
      salt:
        description: 加盐
        type: string
      email:
        description: 邮箱
        type: string
      phone:
        description: 手机号
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      roleIds:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
      updateTime:
              description: 更新时间
              type: string
              format: date-time
...
  SysRoleInfo:
    type: object
    properties:
      id:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      roleDesc:
        description: 角色描述
        type: string
      list:
        description: 资源ID列表
        type: array
        items:
          type: integer
          format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
      updateTime:
        description: 更新时间
        type: string
        format: date-time
...